AUC-Thesis-DT-Physical/RemoteDrivingDashboard-master/requirements.txt
Django==1.11.20
django-cors-headers==3.2.1
eventlet==0.21.0
gevent_socketio==0.3.6
mysql_connector_repackaged==0.3.1
PyMySQL==1.0.2
requests==2.23.0
python-socketio==5.4.1
pandas
tsod
scikit-learn

websocket-client==1.2.1

mysql-connector
mysql-connector-python




